split ydb/library/workload to some separate library
