+++ ydb/core/load_test/ya.make
@@ -16,7 +16,9 @@ PEERDIR(
     ydb/core/kqp/rm_service
     ydb/core/tx/columnshard
     ydb/core/tx/datashard
-    ydb/library/workload
+    ydb/library/workload/abstract
+    ydb/library/workload/kv
+    ydb/library/workload/stock
     ydb/public/lib/base
     ydb/public/lib/operation_id
     ydb/public/sdk/cpp/client/ydb_proto
