+++ ydb/library/workload/kv/registrar.cpp
@@ -1,0 +1,8 @@
+#include "kv.h"
+#include <ydb/library/workload/abstract/workload_factory.h>
+
+namespace NYdbWorkload {
+
+TWorkloadFactory::TRegistrator<TKvWorkloadParams> KvRegistrar("kv");
+
+}
