+++ ydb/library/workload/stock/registrar.cpp
@@ -1,0 +1,8 @@
+#include "stock.h"
+#include <ydb/library/workload/abstract/workload_factory.h>
+
+namespace NYdbWorkload {
+
+TWorkloadFactory::TRegistrator<TStockWorkloadParams> StockRegistrar("stock");
+
+}
